Changes Thomas
Changed the GUI.
There are now two frames, products have buttons with images, some
products are bound to keys, product are first registered to a shopping
list before buying.

The products are now read from the csv database file, which is
automatically created every month. The template file is
Streeplijst_0000-00.csv. By changing the products in this file, the new
products will automatically appear the next month when a new database is
generated. If no image is supplied with the same name as the product, a
placeholder image will be used.

Essential files such as the template database, are checked at startup.

Added a setup file for py2exe which can be run from within Eclipse (if
py2exe is installed to python) to create a Windows executable.